review feedback and comments for next refactoring steps
