+++ src/cubi_tk/sodar/lz_create.py
@@ -91,6 +91,7 @@ class CreateLandingZoneCommand:
         if existing_lzs and self.args.unless_exists:
             lz = existing_lzs[-1]
         else:
+            #TODO: make sure assay_uuid is not none
             lz = api.landingzone.create(
                 sodar_url=self.args.sodar_server_url,
                 sodar_api_token=self.args.sodar_api_token,
